Add files via upload
This code works with the Globe Gizmo. It allows the user to physically move around the globe and find out about the history, culture, people, climate, geography of different cities. The Globe talks to the user and reads out wikipedia pages, whilst simultaneously playing a relevant youtube video for visual effect. The globe promotes the spontaneity of travel by then allowing the user to search for the cheapest flights between cities they have moved between, for a date chosen by them. The aim of the Globe is to excite people and to inspire them to travel!